requirements.txt
torch
torchvision
torchaudio
albumentations
numpy
matplotlib
opencv-python
Pillow
scikit-learn  
torch
torchvision
Pillow
Flask
# updatd ine